% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_trials.R
\name{auc_trials}
\alias{auc_trials}
\title{Calculate area under curve}
\usage{
auc_trials(dataframe, trials, time.range)
}
\arguments{
\item{dataframe}{a fluoR-format dataframe}

\item{trials}{trial numbers to find AUC for}

\item{time.range}{range of time to find AUC for}
}
\value{
a named vector of trial AUCs
}
\description{
`auc_trials` calculates area under curve for multiple
trials using trapezoidal integration
}
\examples{
### What is the AUC for the first 2 seconds of post-onset
### for the first 5 trials?
df <- format_data(GCaMP)
trials <- 1:5
time.range <- c(0,2)

auc_trials(dataframe = df, time.range = time.range, trials = trials)
}
