% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_transform.R
\name{baseline_transform}
\alias{baseline_transform}
\title{Transform trial values relative to baseline}
\usage{
baseline_transform(dataframe, trials, baseline.times, type = "z_standard")
}
\arguments{
\item{dataframe}{a fluoR-format dataframe}

\item{trials}{trial numbers to transform}

\item{baseline.times}{a range of values specifying a baseline period}

\item{type}{type of transformation to use on dataframe input}
}
\value{
a dataframe of transformed values for the specified trials
}
\description{
`baseline_transform` transforms a fluoR dataframe into change
from a specified baseline period
}
\examples{
### Format data frame
df <- format_data(GCaMP)

### Standardize variables relative to baseline using z-scores
baseline_transform(dataframe = df, trials = 1:5, baseline.times = c(-4,0),
type = 'z_standard')
}
