% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_between_peaks.R
\name{distance_between_peaks}
\alias{distance_between_peaks}
\title{Find distance between peaks}
\usage{
distance_between_peaks(dataframe, trial, time.range, n.points = 3)
}
\arguments{
\item{dataframe}{a fluoR-format dataframe}

\item{trial}{trial number for input}

\item{time.range}{range of time to subset}

\item{n.points}{the number of decreasing data points on left and right of
inflection point required to be considered a "peak"}
}
\value{
a dataframe of information on peaks and differences between them
}
\description{
`distance_between_peaks` finds the distance between wave peaks
}
\examples{
### Format data frame
df <- format_data(GCaMP)

### Find distance between peaks for 4s post-stimulus onset for trial 2
distance_between_peaks(dataframe = df, trial = 2, time.range = c(0, 4))
}
