% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflect_points.R
\name{inflect_points}
\alias{inflect_points}
\title{Find inflection points}
\usage{
inflect_points(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
vector of inflection points (-2 or 2) with 0's filling the
points in-between
}
\description{
`inflect_points` finds inflection points of waveform data
}
\examples{
### Format data frame
df <- format_data(GCaMP)

### Find inflection points for 1st trial
inflect_points(df$Trial1)
}
