% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflect_points_trials.R
\name{inflect_points_trials}
\alias{inflect_points_trials}
\title{Find inflection points for multiple trials}
\usage{
inflect_points_trials(dataframe, trials)
}
\arguments{
\item{dataframe}{a fluoR-format dataframe}

\item{trials}{trial numbers to find inflection points for}
}
\value{
a dataframe of timestamps and corresponding inflection points for
each specified trial
}
\description{
`inflect_points_trials` finds inflection points of waveform
data for a range of trials
}
\examples{
### Format data frame
df <- format_data(GCaMP)

### Find inflection points for first 4 trials
inflect_points_trials(dataframe = df, trials = c(1,2,3,4))
}
