% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_change.R
\name{percent_change}
\alias{percent_change}
\title{Calculate percent change}
\usage{
percent_change(xvals, base.val)
}
\arguments{
\item{xvals}{values to be transformed}

\item{base.val}{a baseline value for percent change to be scaled by}
}
\value{
a numeric vector of percent changes from baseline value
}
\description{
`percent_change` transforms a numeric vector into percent
change from a base value
}
\examples{
### Format data frame
df <- format_data(GCaMP)

base.vals <- df$Trial1[df$Time <= 0] # extract pre-event vector (baseline)
base.mean <- mean(base.vals)

percent_change(xvals = df$Trial1, base.val = base.mean)
}
