% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trials.R
\name{plot_trials}
\alias{plot_trials}
\title{Plot range of trials}
\usage{
plot_trials(dataframe, trials, time.range = NULL)
}
\arguments{
\item{dataframe}{a fluoR-format dataframe}

\item{trials}{trial numbers to be plotted}

\item{time.range}{range of time to be plotted; default keeps full range}
}
\value{
a line graph of user-specified trials
}
\description{
`plot_trials` plots trials specified by the user
}
\examples{
### Format data frame
df <- format_data(GCaMP)

### Plot 2 seconds before and 2 seconds after event onset for trials 1-4
my.trials <- c(1, 2, 3, 4)
start.stop <- c(-2, 2)

plot_trials(dataframe = df, trials = my.trials, time.range = start.stop)
}
