% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_trials.R
\name{summarize_trials}
\alias{summarize_trials}
\title{Summarize data across trials}
\usage{
summarize_trials(dataframe, trials, summary.type = "mean")
}
\arguments{
\item{dataframe}{a fluoR-format dataframe}

\item{trials}{a vector of trials to summarize across}

\item{summary.type}{the type of summary statistic to compute}
}
\value{
a dataframe of timestamps and summarized trial values
}
\description{
`summarize_trials` performs the specified summary statistic
across trials for all applicable timestamps
}
\examples{
### Format data frame
df <- format_data(GCaMP)

### Average across trials 1, 2, 3, and 4
trial.vec <- c(1, 2, 3, 4)
type <- 'mean'

summarize_trials(dataframe = df, trials = trial.vec, summary.type = type)
}
